% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_scanner_funcs.R
\name{change.two.parameters}
\alias{change.two.parameters}
\title{Takes sets of two params to be changed and two param sets to be held constant, and selects all the param sets (being changed) which satisfy the target pause and body levels}
\usage{
change.two.parameters(
  func.density = density.prime.kinit.krel,
  p.change = pause.change,
  b.change = body.change,
  baseline.pause = out.pause,
  baseline.body = out.body,
  param1 = kinit.vec,
  param2 = krel.vec,
  params = c("kinit", "krel"),
  constant.param1.vec = kpre.vec,
  constant.param2.vec = kelong.vec,
  perc = 0.05,
  fold.query = 5,
  sample.size = 10000,
  ...
)
}
\arguments{
\item{func.density}{a function which defines the compartment model and accomodates two of the param sets held constant and two params which are changed.}

\item{p.change}{change in pause density}

\item{b.change}{change in body density}

\item{baseline.pause}{a vector containing baseline pause densities}

\item{baseline.body}{a vector containing baseline body densities}

\item{param1}{a vector containing individual realisations of first param to be changed}

\item{param2}{a vector containing individual realisations of second param to be changed}

\item{params}{a string vector naming the two params being changed}

\item{constant.param1.vec}{a vector containing realisations of the first param to be kept constant}

\item{constant.param2.vec}{a vector containing realisations of the second param to be kept constant}

\item{perc}{percent threshold for matching targeted and realized values}

\item{fold.query}{the params to be changed will be sweeped from [1 / fold.query, fold.query] * original param vals}

\item{sample.size}{the number of samples randomly choosen from the sample space}
}
\value{
a list containing two sets of modified parameters, i.e; param1 and param2 sets which satisify change in pause region (baseline.pause*p.change) and gene body (baseline.body*b.change) in the perc threshold provided
}
\description{
Takes sets of two params to be changed and two param sets to be held constant, and selects all the param sets (being changed) which satisfy the target pause and body levels
}
\examples{
#this function can be called separately. It is invoked within parallel.change.two.param.helper() where details are hidden from the user
change.two.parameters(func.density = density.prime.kpre.kelong,
                                 p.change = 1.15,
                                 b.change = 0.95,
                                 baseline.pause = out.pause,
                                 baseline.body = out.body,
                                 param1 = kpre.vec,
                                 param2 = kelong.vec,
                                 params = c("kpre", "kelong"),
                                 constant.param1.vec = kinit.vec,
                                 constant.param2.vec = krel.vec,
                                 perc = 0.05, fold.query = 3, sample.size = 10000)
}
