% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_scanner_funcs.R
\name{density.prime.kpre.krel}
\alias{density.prime.kpre.krel}
\title{an implementation of compartment model where kpre and krel are being modified, and kinit and kelong are held constant. (To be passed to ode() solver in package deSolve(). )}
\usage{
density.prime.kpre.krel(
  t,
  initial.state,
  params = params,
  constant.param1,
  constant.param2
)
}
\arguments{
\item{t}{series of time points where this function is used to evaluate the change in variables}

\item{initial.state}{the values of variables being used to start the model evolution}

\item{params}{containing values of rates being changed, i.e;  kpre and krel}

\item{constant.param1}{the first param not being modified, here kinit}

\item{constant.param2}{the second param not being modified, here kelong}
}
\value{
two values corresponding to pause and body densities
}
\description{
an implementation of compartment model where kpre and krel are being modified, and kinit and kelong are held constant. (To be passed to ode() solver in package deSolve(). )
}
