% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_traces.R
\name{density.prime}
\alias{density.prime}
\title{Compartment model adapted from our G&D paper (doi:10.1101/gad.328237.119):
Declare differential equations as function
dP/dt = kinit - (kpre +krel)p
dB/dt = krel * p - kelong * b
P is the first dependent variable, promoter density; dP is its derivative wrt time
B is the second dependent variable, body density; dB is its derivative wrt time}
\usage{
density.prime(t, initial.state, params = params)
}
\arguments{
\item{t}{time points to run the ODE model}

\item{initial.state}{a list containing initial values of the variables being evolved.}

\item{params}{list of other variables (rates) being used in the model.}
}
\value{
this function is called within ode function, and returns pause body with values in three body compartments.
}
\description{
Compartment model adapted from our G&D paper (doi:10.1101/gad.328237.119):
Declare differential equations as function
dP/dt = kinit - (kpre +krel)p
dB/dt = krel * p - kelong * b
P is the first dependent variable, promoter density; dP is its derivative wrt time
B is the second dependent variable, body density; dB is its derivative wrt time
}
\examples{
library(deSolve)
initial.state = c(P = 1, B1 = 0.01)
parms = c(kinit = 0.1, krel = 0.1, kpre = 0.1, kelong = 50)
t <- seq(from = 0.01, to = 100.01, by = 10)
y <- ode(y = initital.state, times = t, func = density.prime, parms = parms)
}
