% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_composites_lattice.R
\name{dynamic.pro.profile}
\alias{dynamic.pro.profile}
\title{uses [find.body.param()] and [get.pro.waveform()] to generate PRO-seq waveform based on pause and gene body values provided in the input.}
\usage{
dynamic.pro.profile(
  input,
  tau = 20,
  min.pk = 0,
  max.pk = 1,
  dpk = 0.001,
  gene.len = 1000,
  filename = "dynamic_pro_model_density"
)
}
\arguments{
\item{input}{a data farme containing Body and Pause columns with corresponding values}

\item{tau}{a parameter which acts as a "time constant" for the exponential function fitting in find.body.param}

\item{min.pk}{minimum peak value - to be passed to [find.body.param()]}

\item{dpk}{threshold for implicit solution - to be passed to [find.body.param()]}

\item{filename}{name of the file to which the plot is saved}

\item{max.px}{maximum peak value - to be passed to [find.body.param()]}

\item{gen.len}{length of gene i.e; x-range for the desired PRO-seq profile}
}
\description{
uses [find.body.param()] and [get.pro.waveform()] to generate PRO-seq waveform based on pause and gene body values provided in the input.
}
