% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_traces.R
\name{find.body.param}
\alias{find.body.param}
\title{code adapted from our G&D paper (doi:10.1101/gad.328237.119):
to visualize it in a composite profile form
function for finding the gene body parameter}
\usage{
find.body.param(
  bpeak = NULL,
  tau = NULL,
  pausepeak = NULL,
  min.pk = 0,
  max.pk = 1,
  dpk = 0.001
)
}
\arguments{
\item{bpeak}{desired gene body level}

\item{tau}{exponential decay constant}

\item{pausepeak}{desired paused region level}

\item{min.pk}{minimal level for gene body peak}

\item{max.pk}{maximal level for gene body peak}

\item{dpk}{resolution for the implicit solution}
}
\value{
the desired body parameter
}
\description{
code adapted from our G&D paper (doi:10.1101/gad.328237.119):
to visualize it in a composite profile form
function for finding the gene body parameter
}
\examples{
body.param <- find.body.param(bpeak=0.08, tau=20, pausepeak=0.25, min.pk=0, max.pk=1, dpk=.001)
}
