% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_workflow_funcs_v2.R
\name{find.pause.regions}
\alias{find.pause.regions}
\title{Get start, end coordinates of pause window and gene body for each gene using the bigWig files representing all conditions x replicates. The start and end coordinates are treated as transcription start site (TSS) and transcription termination site (TTS) respectively. The pause window has a 50bp size around the peak signal. The gene body starts from pause window end and ends at TTS.}
\usage{
find.pause.regions(bed.input, combined.plus.bw, combined.minus.bw)
}
\arguments{
\item{bed.input}{A bed6 file containing genes of interest, where start and end are transcription start site and transciption termination site respectively.}

\item{combined.plus.bw}{bigWig file for the plus strand data}

\item{combined.minus.bw}{bigWig file for the minus strand data}
}
\value{
A bed6 file containing start, end coordinates of pause window (50bp in size), gene body () starting from seventh column.
}
\description{
Get start, end coordinates of pause window and gene body for each gene using the bigWig files representing all conditions x replicates. The start and end coordinates are treated as transcription start site (TSS) and transcription termination site (TTS) respectively. The pause window has a 50bp size around the peak signal. The gene body starts from pause window end and ends at TTS.
}
\examples{
master.pTA.coords.file = 'mastercoords.bed' # bed file created from primary Transcript Annotation.
pTA <- read.table(master.pTA.coords.file)
df.pause.body <- find.pause.regions(pTA, bw.plus, bw.minus) # bw.* files are corresponding bigWig files.
}
