% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_workflow_funcs_v2.R
\name{generate.composite.df}
\alias{generate.composite.df}
\title{creates composite signal starting from -upstream to +downstream based upon bigWig files passed per condtion.}
\usage{
generate.composite.df(
  df.input,
  cond1.plus.bw,
  cond1.minus.bw,
  cond1 = "cond1",
  cond2.plus.bw,
  cond2.minus.bw,
  cond2 = "cond2",
  upstream = 350,
  downstream = 1400,
  roll.avg = 10,
  step = 5
)
}
\arguments{
\item{df.input}{a bed6 input file - simular to output of [find.pause.regions()] - containing pause.start and pause.end coordinates at seventh and eighth columns.}

\item{cond1.plus.bw}{bigWig file for plus strand data in cond1 (baseline).}

\item{cond1.minus.bw}{bigWig file for plus strand data in cond1 (baseline).}

\item{cond1}{name of cond1 (baseline) to be used as to fill cell values under `cond` column in the returned object,}

\item{cond2.plus.bw}{bigWig file for plus strand data in cond2 (cond2 being the condition of interest being compared against baseline)}

\item{cond2.minus.bw}{bigWig file for minus strand data in cond2 (cond2 being the condition of interest being compared against baseline)}

\item{cond2}{name of the condition which is being compared against baseline}

\item{upstream}{the number of bases upstream from pause summit (default 350)}

\item{downstream}{the number of bases downstream from pause summit (default 1400)}

\item{step}{a factor used to define the ends of genes by (step * roll.avg)/2}

\item{roll, avg}{the rolling size used in rollmean() to calculate the composites (default 10)}
}
\description{
creates composite signal starting from -upstream to +downstream based upon bigWig files passed per condtion.
}
