% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_traces.R
\name{get.pro.waveform}
\alias{get.pro.waveform}
\title{function to get the PRO waveform}
\usage{
get.pro.waveform(
  bpeak = NULL,
  pausepeak = NULL,
  bodypeak = NULL,
  bp.seq = NULL,
  tau = NULL
)
}
\arguments{
\item{bpeak}{desired gene body level}

\item{pausepeak}{desired pause region level}

\item{bodypeak}{body peak value to obtain a level of bpeak}

\item{bp.seq}{base pair sequence}

\item{tau}{exponential decay constant}
}
\value{
a list containing a vector with simulated PRO-seq singal and an object containing the simulated and requested signal
}
\description{
function to get the PRO waveform
}
\examples{
x <- get.pro.waveform(pausepeak=0.25, bpeak=0.02, bp.seq=seq(0,999),tau=20)
}
