% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_workflow_funcs_v2.R
\name{merge.pbody.deseq2}
\alias{merge.pbody.deseq2}
\title{a helper function to merge two data frames on gene names. ASSUMES that both data frames have rownames set to gene names.}
\usage{
\method{merge}{pbody.deseq2}(df.pause.body, deseq2.df)
}
\arguments{
\item{df.pause.body}{a data frame containing start and end coordinate for pause windows and gene body. Ideally, an output of [find.pause.regions()]}

\item{deseq2.df}{a data object created from running DESeq2 workflow and containing response status of genes.}
}
\value{
an object merged on gene names having columns in order of df.pause.body followed by deseq2.df
}
\description{
a helper function to merge two data frames on gene names. ASSUMES that both data frames have rownames set to gene names.
}
