% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_scanner_funcs.R
\name{parallel.change.two.param.helper}
\alias{parallel.change.two.param.helper}
\title{a general function to call [change.two.parameters()] given the condition name and the objects associated with scanned parameters}
\usage{
\method{parallel}{change.two.param.helper}(
  cond1 = "cond1",
  cond2 = "cond2",
  direction = "activated",
  output.param.scan,
  x.fast99.obj,
  perc = 0.02,
  pause.index.baseline,
  pause.sum.change,
  body.change,
  foldquery = 2,
  samplesize = 1000
)
}
\arguments{
\item{cond1}{name of baseline conditon}

\item{cond2}{name of condition which is being compared with the baseline}

\item{direction}{nature of the change in the data being compared between condition of interest and the baseline. Say, "activated"}

\item{output.param.scan}{an object containing the model output from param scans.}

\item{x.fast99.obj}{an object created from using rfast99() function from pksensi package}

\item{perc}{percentage threshold used while matching simulated values with the target values}

\item{pause.index.baseline}{the pause index in the baseline condition}

\item{pause.sum.change}{the change in pause sum between conditon of interest and baseline}

\item{body.change}{the change in body signal between condition of interest and baseline}

\item{foldquery}{the param scan will be done with values ranging between [1/foldquery, foldquery] * original param values}

\item{samplesize}{the number of samples randomly selected from the param space defined by foldquery.}
}
\description{
a general function to call [change.two.parameters()] given the condition name and the objects associated with scanned parameters
}
\examples{
# see paramFittingCompartmentModel.R for a description of calling this function
}
