% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_workflow_funcs_v2.R
\name{plot.bw.plots}
\alias{plot.bw.plots}
\title{creates a plot containing box and whisker plot, violin plot and scatter plot of the change in pause indices between a condition of interest and a baseline.}
\usage{
\method{plot}{bw.plots}(input.df, cond1 = "cond1", cond2 = "cond2", color = "lightblue")
}
\arguments{
\item{input.df}{an object containing pause indices for each gene of interest (oer condition,) Ideally, an output of [total.region.density()]}

\item{cond1}{name of the baseline condition}

\item{cond2}{nqme of the condition which is being compared against baseline}

\item{color}{color to be used in the violin plot}
}
\description{
creates a plot containing box and whisker plot, violin plot and scatter plot of the change in pause indices between a condition of interest and a baseline.
}
