% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_composites_lattice.R
\name{plot.changes.wrt}
\alias{plot.changes.wrt}
\title{plot gene body and pause region density based on the input}
\usage{
\method{plot}{changes.wrt}(input, filename = "dynamic_pro_model")
}
\arguments{
\item{input}{contains iteration, body and pause columns for values for plotting}

\item{filename}{name of the file to which the plot is saved}
}
\description{
plot gene body and pause region density based on the input
}
