% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_workflow_funcs_v2.R
\name{plot.composites}
\alias{plot.composites}
\title{uses output of [generate.composite.df()] to generate plot of composite signal.}
\usage{
\method{plot}{composites}(
  dat,
  fact = "GR",
  comp = "20.v.0",
  summit = "TSS",
  y.low = 0,
  y.high = NULL,
  col.lines = c(rgb(0, 0, 1, 1/2), rgb(1, 0, 0, 1/2), rgb(0.1, 0.5, 0.05, 1/2), rgb(0,
    0, 0, 1/2), rgb(1/2, 0, 1/2, 1/2), rgb(0, 1/2, 1/2, 1/2), rgb(1/2, 1/2, 0, 1/2))
)
}
\arguments{
\item{dat}{output of [generate.composite.df()]}

\item{fact}{name of the factor in the condition being compared against}

\item{comp}{a small string to describe the condition being made}

\item{summit}{a string which will be used as a xlabel - distance from <summit>}

\item{y.low}{the lower limit of the y axis}

\item{y.high}{the max limit of the y axis. If not provided, it estimates it using the maximum estimate present in input data}

\item{col.lines}{colors to be used for drawing the composites}
}
\value{
NULl
}
\description{
uses output of [generate.composite.df()] to generate plot of composite signal.
}
