% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_composites_lattice.R
\name{plot.pro.simulation.composites.2}
\alias{plot.pro.simulation.composites.2}
\title{plots PRO-seq waveform of two conditions present in the input.}
\usage{
\method{plot}{pro.simulation.composites.2}(
  input,
  pause.offset = 0,
  filename = "dynamic_pro_model",
  ylim = c(0, 0.0082),
  trace.col = c("grey50", "#6aa3ce")
)
}
\arguments{
\item{input}{a dataframe where first column is condition (group), second column is y.val, third column is x.val.}

\item{pause.offset}{an offset to place vertical guide lines at [0 - pause.offset, 50 - pause.offset]}

\item{filename}{name of the file to which the plot will be saved}

\item{ylim}{two values to set limits of the y.axis}

\item{trace.col}{two set of colors used for plotting PRO-seq signal. Color indices will be mapped in alphabetical order of condition names.}
}
\description{
plots PRO-seq waveform of two conditions present in the input.
}
