% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_composites_lattice.R
\name{plot.pro.simulation.composites}
\alias{plot.pro.simulation.composites}
\title{create a plot of (simulated) composites given input containing y-values against x-values}
\usage{
\method{plot}{pro.simulation.composites}(input, filename = "dynamic_pro_model")
}
\arguments{
\item{input}{a data frame with three columns in this order - index, y.val and x.val. y.val is plotted against x.val within the function}

\item{filename}{this name is used to save the plot}
}
\description{
create a plot of (simulated) composites given input containing y-values against x-values
}
