% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_composites_lattice.R
\name{plot.simulated.composites.helper}
\alias{plot.simulated.composites.helper}
\title{a helper function to plot PRO-seq waveform derived after simulations. Used in the script plot_simulation_composites.R. Uses data from changing knit and krel and arranges code present from page 40 of compartment modeling vignette.}
\usage{
\method{plot}{simulated.composites.helper}(
  input.data.identifier = "activated.condnvsbaseline",
  output.plot.name = "model_condnvsbaseline_composite",
  baseline.condn.name = "control",
  condn.name = "condn",
  seq.end = 0.1,
  by.step = 1e-04,
  tau = 10,
  dpk = 1e-05,
  dataFolder = "../data"
)
}
\arguments{
\item{input.data.identifier}{a string which is used to identify data objects that are read. Ideally, the data objects were generated using [parallel.change.two.param.helper()] called in paramFittingCompartmentModel.R}

\item{output.plot.name}{name of the file used for saving}

\item{condn.name}{name of the condition which is being compared against baseline}

\item{seq.end}{end of the sequence or range of the x-axis}

\item{by.step}{step size of going from 0 to seq.end}

\item{tau}{a "time-constant" like parameter passed to [find.body.param()]}

\item{dpk}{threshold for the implicit solution}

\item{baseline.cond.name}{name of the baseline condition}
}
\description{
a helper function to plot PRO-seq waveform derived after simulations. Used in the script plot_simulation_composites.R. Uses data from changing knit and krel and arranges code present from page 40 of compartment modeling vignette.
}
