% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_scanner_funcs.R
\name{plot.two.parameter.bw}
\alias{plot.two.parameter.bw}
\title{plots fold change for new parameters (obtained from param scan) w.r.t to original parameters.}
\usage{
\method{plot}{two.parameter.bw}(factor.param1.param2, factor = "GR", y.lab, y.lim = NULL)
}
\arguments{
\item{factor.param1.param2}{an object created from [two.parameter.bw.plot.lattice()] containing three columns fold.change, (modified) rates and description of constant params.}

\item{factor}{name of the factor/experiment which will be used in the filnename as <factor>_change_two_parameters.pdf}

\item{y.lab}{label for y-axis}

\item{y.lim}{two values giving the lower and upper limit of y-axis. If not provided, the upper limit is calculated using the data and lower limit is set to zero.}
}
\description{
plots fold change for new parameters (obtained from param scan) w.r.t to original parameters.
}
