% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_composites_lattice.R
\name{pro.integrated.peak}
\alias{pro.integrated.peak}
\title{takes an input dataframe with Pause and Body signals and returns a PRO-seq waveform with the x-coordinate.}
\usage{
pro.integrated.peak(
  input,
  tau = 20,
  min.pk = 0,
  max.pk = 1,
  dpk = 0.001,
  gene.len = 2000,
  pause.height,
  time.char = "0 min"
)
}
\arguments{
\item{input}{a dataframe containing iteration, pause and body signal}

\item{tau}{a "time constant" parameter passed to [find.body.param()]}

\item{min.pk}{minimum of the peak value - to be passed to [find.body.param()]}

\item{max.pk}{maximum of the peak value - to be passed to [find.body.param()]}

\item{dpk}{threshold for implicit solution - to be passed to [find.body.param()]}

\item{gene.len}{range of the x-axis to be used for plotting.}

\item{pause.height}{height of the pause summit.}

\item{time.char}{character which describes the condition}
}
\value{
a dataframe with columns iteration, PRO-seq signal, x-coordinate
}
\description{
takes an input dataframe with Pause and Body signals and returns a PRO-seq waveform with the x-coordinate.
}
