% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pause_workflow_funcs_v2.R
\name{total.region.density}
\alias{total.region.density}
\title{appends "pause sum" and "body average" columns to a bed6 object created by [find.pause.regions()]. It also appends "pause index" column which is the ratio of "pause sum" and "body average".}
\usage{
total.region.density(
  df.input,
  cond1.plus.bw,
  cond1.minus.bw,
  cond1 = "cond1",
  cond2.plus.bw,
  cond2.minus.bw,
  cond2 = "cond2"
)
}
\arguments{
\item{df.input}{bed6 file created by [find.pause.regions()] containing the start and end of pause window and gene body for genes of interest.}

\item{cond1.plus.bw}{bigWig file containing data from plus strand in cond1 ("cond1" is baseline.)}

\item{cond1.minus.bw}{bigWig file containing data from minus strand in cond1 ("cond1" is baseline.)}

\item{cond1}{name of the condition (baseline)}

\item{cond2.plus.bw}{bigWig file containing data from plus strand in cond2 ("cond1" is the condition of interest being compared against the baseline.)}

\item{cond2.minus.bw}{bigWig file containing data from minus strand in cond2 ("cond1" is the condition of interest being compared against the baseline.)}

\item{cond2}{name of the condition of interest}
}
\value{
the input bed6 object appended with columns containing pause sum, body average, pause index for each gene present in the input file.
}
\description{
appends "pause sum" and "body average" columns to a bed6 object created by [find.pause.regions()]. It also appends "pause index" column which is the ratio of "pause sum" and "body average".
}
\examples{
see documentation of [find.pause.regions()] to see how bed.object was created.
df.input = find.pause.regions(bed.object, bw.plus, bw.minus)
df.region.density <- total.region.density(df.input, cond1.plus.bw, cond1.minus.bw, "cond1",
                                           cond2.plus.bw, cond2.minus.bw, "cond2")
}
