% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_scanner_funcs.R
\name{two.parameter.bw.plot.lattice}
\alias{two.parameter.bw.plot.lattice}
\title{calculates fold change between given param list and original values and returns a data frame}
\usage{
two.parameter.bw.plot.lattice(
  param1.param2.lists,
  param1.vec,
  param2.vec,
  params = c("Kinit", "Krel"),
  constant.params = c("Kelong", "Kpre")
)
}
\arguments{
\item{param1.vec}{original values of param1}

\item{param2.vec}{original values of param2}

\item{params}{names of params which were modified i.e; the identity of param1, param2}

\item{constant.params}{names of parameters which were held constant}

\item{param1.param2.list}{a list containing the derived param values from running [change.two.parameters()]}
}
\value{
data frame containing three columns - fold.change, (modified) rates and description of constant params. The description in 3rd column is constant_<param1>_<param2>
}
\description{
calculates fold change between given param list and original values and returns a data frame
}
\examples{
#the function was taken from page 28 of compartment modeling vignette 
#see ./callers/plotParamSets.R for how these functions were called. The "two param" lists were created by parallel.change.two.param.helper()
}
